Run connection reset tests not on Windows
